@@ -18,10 +18,7 @@ class DocumentRepositoryTest extends TestCase
 {
     use DocumentManagerTrait;
 
-    /**
-     * @var DocumentRepository
-     */
-    private $repository;
+    private DocumentRepository $repository;
 
     /**
      * {@inheritdoc}
