@@ -7,9 +7,9 @@ interface ObjectIteratorInterface extends \Iterator, \Countable
     /**
      * Registers a callable to apply to each element of the iterator.
      *
-     * @param callable $func
+     * @param callable $callable
      *
      * @return $this
      */
-    public function apply(?callable $func = null): self;
+    public function apply(?callable $callable = null): self;
 }
